SmartSearch architecture improvements
